notify: extremely simple notification for Python
This is meant for sending messages to yourself
or a small number of other pre-approved
recipients.

The interface is kept deliberately extremely
simple so that email could easily be switched
out for twilio or something similar.

Sending a message should be as simple as:

    import notify.mail as notify
    notify.send_message("Hello world!")

The program loads your sending credentials _and_
recipients from a configuration file, e.g.
~/.notify.yaml. (Naturally be careful with
permissions on this file.)

Currently some work needs to be done
on "emailification" if we get a naïve string,
e.g. the above string instead of:

"""\
Subject: Hello world!

Hello there
"""