work around awkward namespace issue
The nature of this library means that a lot of
the "sub-modules" will be small wrappers around
libraries previously written for protocols and
services. For instance, both "email" and "twilio"
already exist as module names.

Because absolute imports are still in the
__future__ (including in a lot of code that I
don't control), reusing these names causes
trouble.

The solution chosen in this workaround is to
give all the service modules a prefix ("notify_")
and then introduce a "services" helper object
to perform dynamic imports without writing
the awkward prefix.